Fix incorrect ciphersuite setup for libreswan
ESP does not have a PRF, and GCM has no INTEG/AUTH, so
"aes_gcm256-sha2" is not a valid specifier. That should be
"aes_gcm256"

The aes-sha1 should not be preferred over aes-sha2_512.
The only exception to regular sha strengths, is that for ESP,
sha2_256 should have a lower priority (appear later in the line)
then sha1, due to a kernel bug and android use.